Create a ReservationList component
