+++ src/CurrentReservations.js
@@ -1,86 +1,30 @@
-import React, {Component, PropTypes} from "react";
+import React from "react";
 import {
-    Body,
     Card,
     CardItem,
-    Container,
-    Icon,
-    Left,
-    Right,
     Text,
-    Spinner
 } from "native-base";
-import {View} from "react-native";
-import moment from "moment";
+
+import ReservationList from "./ReservationList";
 
 export default class CurrentReservations extends React.Component {
 
-    static PropTypes = {
-        reservation: PropTypes.array,
-        onItemPressed: PropTypes.function
+    static propTypes = {
+        reservations: React.PropTypes.array,
+        onItemPressed: React.PropTypes.func
     };
 
     render() {
-      return (
-          <Card>
-              <CardItem>
-                  <Text>Current reservations</Text>
-              </CardItem>
-              {this.renderReservations()}
-          </Card>
-      );
-    }
-
-    renderReservations() {
-        const {reservations} = this.props;
-        if (!reservations) {
-            return <CardItem style={styles.itemWSpinner}><Spinner color='blue' /></CardItem>;
-        }
-        return reservations.map(reservation => {
-            if (reservation.status === "A") {
-                return (
-                    <CardItem
-                        button
-                        onPress={() => this.props.onItemPressed(reservation)}
-                        key={reservation.pk}
-                    >
-                        <Left>
-                            <Text style={styles.marginRight}>{reservation.service.name}</Text>
-                        </Left>
-                        <Body>
-                            {this.renderReservationTime(
-                                reservation.from_datetime,
-                                reservation.to_datetime
-                            )}
-                        </Body>
-                        <Right>
-                            <Icon name="arrow-forward" />
-                        </Right>
-                    </CardItem>
-                );
-            }
-        });
-    }
-
-    renderReservationTime(fromDate, toDate) {
-        from = moment(fromDate).format("HH:mm");
-        to = moment(toDate).format("HH:mm");
-        if (from === "00:00" && to == "00:00") {
-            return <Text>All day</Text>
-        }
         return (
-            <Text>From {from} to {to}</Text>
+            <Card>
+                <CardItem>
+                    <Text>Current reservations</Text>
+                </CardItem>
+                <ReservationList 
+                    reservations={this.props.reservations}
+                    onItemPressed={this.props.onItemPressed}
+                />
+            </Card>
         );
     }
 }
-
-
-const styles = {
-    itemWSpinner: {
-        flex: 1,
-        justifyContent: "center"
-    },
-    marginRight: {
-        marginRight: 15
-    }
-};
